gha: don't fail if all cloud provider matrix entries are filtered out
Cloud provider related E2E workflows have been recently updated to
automatically filter out the matrix entries if the corresponding k8s
version is EOL [1,2,3]. However, if that happen, the corresponding
workflow currently fails due to both the empty matrix and the
impossibility of subsequently merging the artifacts. Let's fix this
by checking if the resulting matrix is empty, and explicitly skipping
the subsequent tasks in that case.

[1]: dbcdd7dbe2e9 ("ci: Filter supported versions of AKS")
[2]: 720927534b26 ("ci: Filter supported versions of EKS")
[3]: dd947b3a3830 ("ci: Filter supported versions of GKE.")

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>